Bump version to 0.36-SNAPSHOT; prepare for PR submission
